requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas
numpy
matplotlib
scikit-learn
tqdm
Pillow
open_clip_torch
faiss-cpu
radgraph
modal
